+++ src/interfaces/IVault.sol
@@ -2,11 +2,11 @@
 pragma solidity = 0.8.17;
 
 interface IVault {
-  event Withdraw (uint indexed from, address indexed to, uint amount);
-  event Deposit  (uint indexed id, uint amount);
-  event Move     (uint indexed from, uint indexed to, uint amount);
+    event Withdraw(uint256 indexed from, address indexed to, uint256 amount);
+    event Deposit(uint256 indexed id, uint256 amount);
+    event Move(uint256 indexed from, uint256 indexed to, uint256 amount);
 
-  error StaleData            ();
-  error IncompleteRound      ();
-  error NotVaultManager      ();
+    error StaleData();
+    error IncompleteRound();
+    error NotVaultManager();
 }
