+++ src/interfaces/IVaultManager.sol
@@ -2,22 +2,22 @@
 pragma solidity = 0.8.17;
 
 interface IVaultManager {
-  event Added     (uint indexed id, address indexed vault);
-  event Removed   (uint indexed id, address indexed vault);
-  event MintDyad  (uint indexed id, uint amount, address indexed to);
-  event BurnDyad  (uint indexed id, uint amount, address indexed from);
-  event RedeemDyad(uint indexed id, address indexed vault, uint amount, address indexed to);
-  event Liquidate (uint indexed id, address indexed from, uint indexed to);
+    event Added(uint256 indexed id, address indexed vault);
+    event Removed(uint256 indexed id, address indexed vault);
+    event MintDyad(uint256 indexed id, uint256 amount, address indexed to);
+    event BurnDyad(uint256 indexed id, uint256 amount, address indexed from);
+    event RedeemDyad(uint256 indexed id, address indexed vault, uint256 amount, address indexed to);
+    event Liquidate(uint256 indexed id, address indexed from, uint256 indexed to);
 
-  error NotOwner();
-  error NotLicensed();
-  error OnlyOwner();
-  error VaultNotLicensed();
-  error TooManyVaults();
-  error VaultAlreadyAdded();
-  error VaultHasAssets();
-  error NotDNftVault();
-  error InvalidDNft();
-  error CrTooLow();
-  error CrTooHigh();
+    error NotOwner();
+    error NotLicensed();
+    error OnlyOwner();
+    error VaultNotLicensed();
+    error TooManyVaults();
+    error VaultAlreadyAdded();
+    error VaultHasAssets();
+    error NotDNftVault();
+    error InvalidDNft();
+    error CrTooLow();
+    error CrTooHigh();
 }
